Adding 404 response for non-existant movie ID

+++ src/main/java/tech/joes/Repositories/MovieRepository.java
@@ -6,5 +6,5 @@ import tech.joes.Models.Movie;
 /**
  * Created by joe on 05/04/2017.
  */
-public interface MovieRepository extends CrudRepository<Movie,Long> {
+public interface MovieRepository extends CrudRepository<Movie,Integer> {
 }
